Initial Commit
An article that appeared in the *British Medical Journal* (2010) presented the results of a study conducted by researcher Jeremy Groves, whose objective was to determine whether the type of frame of his bicycle could affect his travel time to work. 

On each of 56 days from mid-January to mid-July 2010, Dr. Groves tossed a coin to decide whether he would ride the 27 miles to work on his carbon frame bicycle that was light in weight, weighing 20.90 pounds, or on his steel framed bicycle that was heavier, weighing 29.75 pounds. He then recorded that commute time **in minutes** for each trip.